got videos listed with a few hitches
The main problem is that if I don't want a bunch of red in the console from youtube then I have to remove my line that embeds the videos. I hope this will only be an issue with the stored data and the way it's loaded. Other than that I'm not happy with the css and I am thinking of using some bootstrap to help me out with that. next is to get the remove button to work :)