Completed POST request
Used uuid and slice method to make an id of six characters. We destructured title and URL from req.body and validated them. Pushed the new video with the ID to the videos data. Finally, wrapped the id inside the curlies in order to see the property name and its value as response.