Level 250 WIP
Used a new useState variable called fetchData and set it as true or false in the MainContainer to rerender the data after a delete has taken place. Made the delete button work for both the client and the server. Since we have the id already in normal numbers format in our database or exampleresponse.json, to make it consistent with IDs that already exist, we take the highest ID number and add 1 to it to get a new number ID. We generate the ID using Math.max() method. After that I fixed the ID type mismatch in server and the client by making them all converted to number from string.